cursorposition in cursor file
