+++ src/cursor.rs
@@ -1,0 +1,46 @@
+/// Represents the position of the cursor relative to words in the buffer.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum CursorPosition {
+    /// The cursor is in the word with the specified index.
+    InWord(usize),
+
+    /// The cursor is on the left edge of the word with the specified index.
+    /// For example: `abc |hi`, where `|` is the cursor.
+    OnWordLeftEdge(usize),
+
+    /// The cursor is on the right edge of the word with the specified index.
+    /// For example: `abc| hi`, where `|` is the cursor.
+    OnWordRightEdge(usize),
+
+    /// The cursor is not in contact with any word. Each `Option<usize>` specifies the index of the
+    /// closest word to the left and right, respectively, or `None` if there is no word on that side.
+    InSpace(Option<usize>, Option<usize>),
+}
+
+impl CursorPosition {
+    pub fn get(cursor: usize, words: &[(usize, usize)]) -> CursorPosition {
+        use CursorPosition::*;
+
+        if words.is_empty() {
+            return InSpace(None, None);
+        } else if cursor == words[0].0 {
+            return OnWordLeftEdge(0);
+        } else if cursor < words[0].0 {
+            return InSpace(None, Some(0));
+        }
+
+        for (i, &(start, end)) in words.iter().enumerate() {
+            if start == cursor {
+                return OnWordLeftEdge(i);
+            } else if end == cursor {
+                return OnWordRightEdge(i);
+            } else if start < cursor && cursor < end {
+                return InWord(i);
+            } else if cursor < start {
+                return InSpace(Some(i - 1), Some(i));
+            }
+        }
+
+        InSpace(Some(words.len() - 1), None)
+    }
+}
